+++ Lineaire_Algebra/Vector.cpp
@@ -1,7 +1,7 @@
 #include "Vector.h"
 #include "Matrix.h"
 
-Vector::Vector(float x, float y, float z) : x{ x }, y{ y }, z{ z }
+Vector::Vector(double x, double y, double z) : x{ x }, y{ y }, z{ z }
 {
 }
 
@@ -59,7 +59,7 @@ void Vector::operator-=(const Vector& other)
 	z -= other.z;
 }
 
-Vector Vector::operator*(const float& other)
+Vector Vector::operator*(const double& other)
 {
 	Vector v = *this;
 	v.x *= other;
@@ -68,14 +68,14 @@ Vector Vector::operator*(const float& other)
 	return v;
 }
 
-void Vector::operator*=(const float& other)
+void Vector::operator*=(const double& other)
 {
 	x *= other;
 	y *= other;
 	z *= other;
 }
 
-Vector Vector::operator/(const float& other)
+Vector Vector::operator/(const double& other)
 {
 	Vector v = *this;
 	v.x /= other;
@@ -84,14 +84,14 @@ Vector Vector::operator/(const float& other)
 	return v;
 }
 
-void Vector::operator/=(const float& other)
+void Vector::operator/=(const double& other)
 {
 	x /= other;
 	y /= other;
 	z /= other;
 }
 
-//Vector operator*(const float& number, Vector& other)
+//Vector operator*(const double& number, Vector& other)
 //{
 //	return Vector(other.x * number, other.y * number, other.z * number);
 //}
@@ -107,17 +107,17 @@ void Vector::draw(const Graphics& graphics) const
 	graphics.drawVector((int)x, (int)y, (int)z);
 }
 
-const float& Vector::getX() const
+const double& Vector::getX() const
 {
 	return x;
 }
 
-const float& Vector::getY() const
+const double& Vector::getY() const
 {
 	return y;
 }
 
-const float& Vector::getZ() const
+const double& Vector::getZ() const
 {
 	return z;
 }
